Updating the Base58Encode function and a few tests.
Updating the Base58Encode function to handle leading zeros (zeroes?), Values that are converted to base58 can have leading zeros at the front, which will be lost when converted to a BigInteger, so we need to count them and convert them to their base58 counterpart, '1'. I also added a functino to convert a base58 string back to a big int.
I could have used fewer lines here but I like the clear steps in the recursive functions.