Ported from .NET Framework to .NET Core, and added BitcoinAddress generation.
Lots of changes here. To start, I ported the project to .NET Core because DriveHappy said to, but also because my test project was already .NET Core. When trying to create the RIPEMD160Managed type, I was getting an error because MICROSOFT REMOVED THIS FOR SOME REASON in later versions of .NET Framework! After being thoroughly upset about this, I ported the project over to .NET Core and found a NuGet package that has the RIPEMD160Managed type defined. So here's a quick summary:
- Added a BitcoinAddressRecord type. It will contain all the information and metadata related to generating a bitcoin address.
- Added a few functions necessary to generate a bitcoin address (p2pkh and p2sh).
- Changed targeting of BitcoinLib.fs from .NET Framework to .NET Core.
- Added two nuget packages, Bouncy Castle to get the secp256k1 elliptic curve, and RIPEMD160 since I need it and MS removed it.
- Added a mega test to test that all the metadata for a bitcoin address was generated correctly.