Updating more functions in Crypto.cs to return result type.
Updating the function Sha256 to return a Result. This causes a cascade of changes upstream, but it's easily handled with the result monad. Had to update a few tests to check for a Result.Error instead of just assuming things are good.
I also moved the GenerateChecksum function into Encoding.cs and renamed it to DoubleSha, since thats what generating a checksum uses.