Wrapping more .NET calls with exception handing.
I'm just bubbling the exception up as an Error string.
- Split the ByteArrayToHex function into two separate functions. It's a little redundant but cleaner than having the bool passed in I think.
- Changed a million functions to use a result monad to unwrap all the Results being returned.
- Added a Zero function to the result computational expression. This is necessary to allow Assert.{x} to be called inside of a result bock.
- Updated all tests to use the result monad where necessary.
- Added a TestHelper file that will hold helper functions for tests. Right now a lot of tests should fail when an Error is returned mid-test, so I wrote a function that will unwrap and Error and Assert Fail on it.

This seems like a lot of work for functions that don't appear to be called in a way that can throw, but in case they are used in the future I'd rather safe guard them now. Runtime exceptions are a pain, and I'd rather the compiler enforce that the consumer handles the Error case from any function they call.